chore: remove unused ui i18n index
