--- packages/ui/src/utils/i18n/index.ts
@@ -1,1 +1,0 @@
-export * from './multilingual';
