+++ packages/ui/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './style';
-export * from './i18n';
